Refactor folers
