+++ src/IdentityServer/Models/ErrorViewModel.cs
@@ -1,0 +1,21 @@
+﻿// Copyright (c) Brock Allen & Dominick Baier. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
+
+using IdentityServer4.Models;
+
+namespace IdentityServer.Models
+{
+    public class ErrorViewModel
+    {
+        public ErrorViewModel()
+        {
+        }
+
+        public ErrorViewModel(string error)
+        {
+            Error = new ErrorMessage { Error = error };
+        }
+
+        public ErrorMessage Error { get; set; }
+    }
+}
