+++ src/IdentityServer/Models/LogoutInputModel.cs
@@ -1,0 +1,10 @@
+﻿// Copyright (c) Brock Allen & Dominick Baier. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
+
+namespace IdentityServer.Models
+{
+    public class LogoutInputModel
+    {
+        public string LogoutId { get; set; }
+    }
+}
