--- src/IdentityServer/Quickstart/Account/LoginInputModel.cs
@@ -1,20 +1,0 @@
-// Copyright (c) Brock Allen & Dominick Baier. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
-
-using System.ComponentModel.DataAnnotations;
-
-namespace IdentityServer
-{
-    public class LoginInputModel
-    {
-        [Required]
-        public string Username { get; set; }
-
-        [Required]
-        public string Password { get; set; }
-
-        public bool RememberLogin { get; set; }
-        public string ReturnUrl { get; set; }
-        public int IdentityTokenLifetime { get; set; }
-    }
-}
