--- src/IdentityServer/Quickstart/Consent/ProcessConsentResult.cs
@@ -1,18 +1,0 @@
-// Copyright (c) Brock Allen & Dominick Baier. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
-
-namespace IdentityServer
-{
-    public class ProcessConsentResult
-    {
-        public bool IsRedirect => RedirectUri != null;
-        public string RedirectUri { get; set; }
-        public string ClientId { get; set; }
-
-        public bool ShowView => ViewModel != null;
-        public ConsentViewModel ViewModel { get; set; }
-
-        public bool HasValidationError => ValidationError != null;
-        public string ValidationError { get; set; }
-    }
-}
