--- src/IdentityServer/Quickstart/Home/ErrorViewModel.cs
@@ -1,21 +1,0 @@
-﻿// Copyright (c) Brock Allen & Dominick Baier. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
-
-using IdentityServer4.Models;
-
-namespace IdentityServer
-{
-    public class ErrorViewModel
-    {
-        public ErrorViewModel()
-        {
-        }
-
-        public ErrorViewModel(string error)
-        {
-            Error = new ErrorMessage { Error = error };
-        }
-
-        public ErrorMessage Error { get; set; }
-    }
-}
