+++ src/IdentityServer/Views/Shared/_ValidationScriptsPartial.cshtml
